+++ config/packages/api_platform.yaml
@@ -1,4 +1,9 @@
 api_platform:
+    swagger:
+        api_keys:
+            JWT:
+                name: Authorization
+                type: header
     title: API Soluval
     version: 1.0.0
     defaults:
