+++ config/packages/lexik_jwt_authentication.yaml
@@ -3,3 +3,8 @@ lexik_jwt_authentication:
     public_key: '%env(resolve:JWT_PUBLIC_KEY)%'
     pass_phrase: '%env(JWT_PASSPHRASE)%'
     token_ttl: 604800
+
+    api_platform:
+        check_path: /api/login_check
+        username_path: name
+        password_path: password
