+++ config/routes.yaml
@@ -20,3 +20,4 @@ user_update_password:
 
 api_login_check:
     path: /api/login_check
+    methods: [POST]
