--- src/Controller/Api/AuthController.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace App\Controller\Api;
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Lexik\Bundle\JWTAuthenticationBundle\Services\JWTTokenManagerInterface;
-use Lexik\Bundle\JWTAuthenticationBundle\Encoder\JWTEncoderInterface;
-
-class AuthController extends AbstractController
-{
-    private $jwtManager;
-    private $jwtEncoder;
-
-    public function __construct(JWTTokenManagerInterface $jwtManager, JWTEncoderInterface $jwtEncoder)
-    {
-        $this->jwtManager = $jwtManager;
-        $this->jwtEncoder = $jwtEncoder;
-    }
-
-    #[Route('/api/login', name: 'api_login', methods: ['POST'])]
-    public function login(UserInterface $user): JsonResponse
-    {
-        $token = $this->jwtManager->create($user);
-        $decodedToken = $this->jwtEncoder->decode($token);
-
-        return $this->json([
-            'token' => $token,
-            'expiration' => date('Y-m-d H:i:s', $decodedToken['exp']),
-        ]);
-    }
-}
