+++ templates/base.html.twig
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+{# <!DOCTYPE html>
 <html lang="fr">
 	<head>
 		<meta charset="UTF-8">
@@ -35,4 +35,4 @@
 				{% endblock %}
 			{% endblock %}
 		</body>
-	</html>
+	</html> #}
